+++ web/packages/studio/src/components/AddColumnPalette/index.tsx
@@ -1,14 +1,11 @@
 // SPDX-FileCopyrightText: Copyright (c) 2025-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import { Flex, Stack, Text, TextInput } from '@nvidia/foundations-react-core';
-import {
-  COLUMN_TYPE_GROUPS,
-  ICON_COLOR_CLASS,
-} from '@studio/components/AddColumnPalette/constants';
+import { Stack, Text, TextInput } from '@nvidia/foundations-react-core';
+import { ColumnTypeGroupSection } from '@studio/components/AddColumnPalette/ColumnTypeGroupSection';
+import { COLUMN_TYPE_GROUPS } from '@studio/components/AddColumnPalette/constants';
 import type {
   AddColumnSelection,
-  ColumnTypeGroup,
   ColumnTypeOption,
 } from '@studio/components/AddColumnPalette/types';
 import { Search } from 'lucide-react';
@@ -17,64 +14,6 @@ import { type FC, useMemo, useState } from 'react';
 /** Matches an option against a lowercased search query (name + description). */
 const matchesQuery = (option: ColumnTypeOption, query: string): boolean =>
   option.label.toLowerCase().includes(query) || option.description.toLowerCase().includes(query);
-
-interface ColumnTypeCardProps {
-  option: ColumnTypeOption;
-  onSelect: (selection: AddColumnSelection) => void;
-}
-
-/**
- * A single column-type option, rendered as a native `<button>` so it is reachable and
- * activatable by keyboard (Tab to focus, Enter/Space to add) with no drag interaction.
- */
-const ColumnTypeCard: FC<ColumnTypeCardProps> = ({ option, onSelect }) => {
-  const { icon: Icon, label, description, color, columnType, samplerType } = option;
-  return (
-    <button
-      type="button"
-      onClick={() => onSelect({ columnType, samplerType })}
-      className="flex cursor-pointer w-full items-center gap-2 rounded-md border border-base bg-surface-raised px-2 py-1.5 text-left transition-colors hover:border-strong hover:bg-surface-hover focus-visible:border-strong focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-[#76b900]"
-    >
-      <Flex
-        align="center"
-        justify="center"
-        className="size-[26px] shrink-0 rounded-sm bg-surface-sunken"
-      >
-        <Icon size={15} className={ICON_COLOR_CLASS[color]} aria-hidden />
-      </Flex>
-      <Stack gap="density-xxs" className="min-w-0">
-        <Text kind="body/semibold/sm" className="truncate text-primary">
-          {label}
-        </Text>
-        <Text kind="body/regular/xs" className="truncate text-secondary">
-          {description}
-        </Text>
-      </Stack>
-    </button>
-  );
-};
-
-interface ColumnTypeGroupSectionProps {
-  group: ColumnTypeGroup;
-  options: ColumnTypeOption[];
-  onSelect: (selection: AddColumnSelection) => void;
-}
-
-/** A labeled group heading (with a count) above its option cards. */
-const ColumnTypeGroupSection: FC<ColumnTypeGroupSectionProps> = ({ group, options, onSelect }) => (
-  <Stack gap="1" className="w-full">
-    <Flex align="center" gap="density-xs">
-      <Text kind="label/bold/xs" className="uppercase tracking-wide text-secondary">
-        {group.label}
-      </Text>
-    </Flex>
-    <Stack gap="1.5" className="w-full">
-      {options.map((option) => (
-        <ColumnTypeCard key={option.id} option={option} onSelect={onSelect} />
-      ))}
-    </Stack>
-  </Stack>
-);
 
 export interface AddColumnPaletteProps {
   /** Called with the chosen column type when an option is activated. */
